requirements.txt
openai>=1.52.0
pydantic>=2.9.0
python-dotenv>=1.0.0
mcp>=1.0.0
httpx>=0.27.0
